+++ src/viewpanel/ViewPanel.tsx
@@ -31,7 +31,7 @@ export const ViewPanel: React.FunctionComponent<IViewPanelProps> = (props: React
 
   if (!metadata || Object.keys(metadata).length === 0) {
     return (
-      <BaseView />
+      <BaseView settings={settings} />
     );
   }